v1.6.8 kind of fixed coeff based hyperparam coll avoidance, Frenet visualization
Coefficient-based hyperparam collision avoidance has been fixed. Or at least kind-of patched. The issue was, that when scaling the clipped corner splines (in the PvX update), because the knots had some numerical deviance from the true desired values, at t = 0 or t = 1 the basis was evaluated to be 0, hence the last or first coefficients after the basis transform were zeroed out. Which caused the problem in collision avoidance.

Frenet based visualization: in the MPC version we now have the ability to plot what the vehicle sees in the Frenet frame and how it plans it's trajectory.