v1.7.4 Obstacle along the way generation + inside collision zone bug fix
Fixed a bug, where collision with obstacle was not detected, due to the fact, that before it only checked, whether any of the corners of the obstacle were inside the danger zone. Now, for example, if the danger zone is smaller, than the obstacle itself, all corners are outside the danger zone. Hence no collision will be detected, even though the circle of the danger zone is inside the polygon, which represents the obstacle.

TODO:
1. need to be able to generate gates and obstacles on one side of the path.
2. need to look at why the formation candidates moove and don't stay at the place, where they were generated.