v1.7.4.2 gate added. Bad results.
Okay. So the reason is probably, that when extrapolating a (which is the normal of the collision avoidance hyperplane, it is extrapolated (due to its "rotational value") in the wrong direction.
Meaning. assume a vehicle, which should avoid an incoming obstacle from below, however, it is slight above its lowest corner. When it approaches it, a will define a perpendicular line ( (vehicle)---> |  [obstacle], where "|" represents the hyperplane). However, before approaching, the hyperplane will probably be tilted a little bit, most likely turned in the way, that is actually good for us (because in the other case, it is likely, that it will go through the obstacle (of course not necessarily, because depending on the value of b, there might be a large gap between them).
However, when extrapolating its spline components, due to its likely rotation (derivative) from “good” to “vertical”, after extrapolation we will get “bad” at the end, because of the “negative/bad directional” derivative.