v1.7.4.7 check functions
We can  now check which equation was infeasible.
We have noticed, that even if the solver says, that it is infeasible, feeding back the result (warm starting it with it's own solution) and starting the optimization again, we often arrive to a successfully solved problem, even though pretty much nothing has changed.