v1.7.5.1 Increased knot intervals for obstacle bspline approximation
At the beginning we generate BSpline trajectories for the corners of the obstacles. With increasing the length of the Frenet path, we have noticed, that the BSpline approximation with 10 knot intervals cannot sufficiently approximate the real movement of the obstacles. As a result, (to be honest, I don't know exactly where), some things went wrong. For example in the case of seed 0, at frame 68-69 one vehicle almost collided...