v1.7.6.4 Event-base MIP working. But not yet 100%, as you might expect : ))
What is also strange, is that if the gate max/min goes to +/-6, then the computation time is around 4.8 seconds.
However if they only extend to +/-2, then the MIP is triggered less number of times. However, the computational cost increases to around 10 seconds. But why?? Less event based triggering is happening so why is the computational time increasing?