v1.7.6.5 bug fix
The problem was, that by rotating the obstacles left-and-right accidentally dx became dy and vica versa. And when taking the mean of these values for dx and dy for example for the gates we have notices, that way-points were generated inside the obstacles. That is because for example in the case of the gates, where dx is much smalle,r than dy, Taking the mean of dy values (which actually also contained dx values because of the bug) the man of dy got smaller. Therefore the vertical extent of the obstacle was virtually smaller and therefore the collision avoidance thought it had more space. Which wasn't true. So now, we also have a time-index for the dxy list, meaning dxy values might be diffrent from time-to-time instance, depending on the rotation of the gate for example.